+++ src/app/features/admin/admin.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AdminRoutingModule } from './admin-routing.module';
+import { AdminComponent } from './admin.component';
+import { FormsModule } from '@angular/forms';
+import { SharedModule } from 'src/app/shared/shared.module';
+
+@NgModule({
+  declarations: [
+    ...AdminRoutingModule.components,
+    AdminComponent
+  ],
+  exports: [
+    ...AdminRoutingModule.components
+  ],
+  imports: [
+    CommonModule,
+    SharedModule,
+    AdminRoutingModule
+  ]
+})
+export class AdminModule { }