+++ src/app/features/admin/components/index.ts
@@ -1,0 +1,4 @@
+export * from './orders/orders.component';
+export * from './product-add/product-add.component';
+export * from './product-edit/product-edit.component';
+export * from './products-manage/products-manage.component';