+++ src/app/shared/components/index.ts
@@ -1,0 +1,2 @@
+export * from './not-found/not-found.component';
+export * from './restricted-access/restricted-access.component';