+++ src/app/shared/components/restricted-access/restricted-access.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'shp-restricted-access',
+  templateUrl: './restricted-access.component.html',
+  styleUrls: ['./restricted-access.component.scss']
+})
+export class RestrictedAccessComponent {
+
+  constructor() { }
+
+}