+++ src/app/shared/directives/index.ts
@@ -1,0 +1,3 @@
+export * from './click-shadow.directive';
+export * from './highlight.directive';
+export * from './no-scroll.directive';