+++ src/app/shp-routing.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { ProductListComponent } from './features/products/components/product-list/product-list.component';
 import { ProductViewComponent } from './features/products/components/product-view/product-view.component';
+import { NotFoundComponent, RestrictedAccessComponent } from './shared/components';
 
 const routes: Routes = [
   {
@@ -19,8 +20,21 @@ const routes: Routes = [
     loadChildren: () => import('./features/cart/cart.module').then(m => m.CartModule)
   },
   {
-    path: '**',
+    path: 'admin',
+    loadChildren: () => import('./features/admin/admin.module').then(m => m.AdminModule)
+  },
+  {
+    path: 'restricted',
+    component: RestrictedAccessComponent,
+  },
+  {
+    path: '',
     redirectTo: 'products-list',
+    pathMatch: 'full',
+  },
+  {
+    path: '**',
+    component: NotFoundComponent,
   },
 ];